reshaping resolver partially done
